nbscreenshot to own function, 'canvas_only' replaes 'with_viewer' kwarg
