+++ napari/_tests/test_viewer.py
@@ -98,28 +98,13 @@ def test_screenshot(viewer_factory):
     data = 20 * np.random.random((10, 4, 2))
     viewer.add_shapes(data)
 
-    # Take screenshot
-    screenshot = viewer.screenshot()
+    # Take screenshot of the image canvas only
+    screenshot = viewer.screenshot(canvas_only=True)
     assert screenshot.ndim == 3
 
     # Take screenshot with the viewer included
-    screenshot = viewer.screenshot(with_viewer=True)
+    screenshot = viewer.screenshot(canvas_only=False)
     assert screenshot.ndim == 3
-
-
-def test_nbscreenshot(viewer_factory):
-    """Test taking a screenshot."""
-    view, viewer = viewer_factory()
-
-    np.random.seed(0)
-    data = np.random.random((10, 15))
-    viewer.add_image(data)
-
-    rich_display_object = viewer.nbscreenshot()
-    assert hasattr(rich_display_object, '_repr_png_')
-    # Trigger method that would run in jupyter notebook cell automatically
-    rich_display_object._repr_png_()
-    assert rich_display_object.image is not None
 
 
 def test_update(viewer_factory):
