+++ napari/_vispy/_tests/test_vispy_multiscale.py
@@ -73,7 +73,7 @@ def test_multiscale_screenshot(viewer_factory):
     # Set canvas size to target amount
     view.view.canvas.size = (800, 600)
 
-    screenshot = viewer.screenshot()
+    screenshot = viewer.screenshot(canvas_only=True)
     center_coord = np.round(np.array(screenshot.shape[:2]) / 2).astype(np.int)
     target_center = np.array([255, 255, 255, 255], dtype='uint8')
     target_edge = np.array([0, 0, 0, 255], dtype='uint8')
@@ -110,7 +110,7 @@ def test_multiscale_screenshot_zoomed(viewer_factory):
     # Check that current level is bottom level of multiscale
     assert viewer.layers[0].data_level == 0
 
-    screenshot = viewer.screenshot()
+    screenshot = viewer.screenshot(canvas_only=True)
     center_coord = np.round(np.array(screenshot.shape[:2]) / 2).astype(np.int)
     target_center = np.array([255, 255, 255, 255], dtype='uint8')
     screen_offset = 3  # Offset is needed as our screenshots have black borders
@@ -142,7 +142,7 @@ def test_image_screenshot_zoomed(viewer_factory):
     view.view.camera.rect = [1000, 1000, 200, 150]
     list(view.layer_to_visual.values())[0].on_draw(None)
 
-    screenshot = viewer.screenshot()
+    screenshot = viewer.screenshot(canvas_only=True)
     center_coord = np.round(np.array(screenshot.shape[:2]) / 2).astype(np.int)
     target_center = np.array([255, 255, 255, 255], dtype='uint8')
     screen_offset = 3  # Offset is needed as our screenshots have black borders
