+++ napari/utils/notebook_display.py
@@ -1,5 +1,24 @@
 from io import BytesIO
 
+__all__ = ['nbscreenshot']
+
+
+def nbscreenshot(viewer, canvas_only=False):
+    """Display napari screenshot in the jupyter notebook.
+
+    Parameters
+    ----------
+    with_viewer : bool, optional
+        If True includes the napari viewer frame in the screenshot,
+        otherwise just includes the canvas. By default, True.
+
+    Returns
+    -------
+    napari.utils.notebook_display.NotebookScreenshot
+        Napari screenshot rendered as rich display in the jupyter notebook.
+    """
+    return NotebookScreenshot(viewer, canvas_only=canvas_only)
+
 
 class NotebookScreenshot:
     """Display napari screenshot in the jupyter notebook.
@@ -23,19 +42,20 @@ class NotebookScreenshot:
     ```
     """
 
-    def __init__(self, viewer, with_viewer=True):
+    def __init__(self, viewer, canvas_only=False):
         """Initalize screenshot object.
 
         Parameters
         ----------
         viewer : napari.Viewer
             The napari viewer
-        with_viewer : bool, optional
-            If True includes the napari viewer frame in the screenshot,
-            otherwise just includes the canvas. By default, True.
+        canvas_only : bool, optional
+            If False include the napari viewer frame in the screenshot,
+            and if True then take screenshot of just the image display canvas.
+            By default, False.
         """
         self.viewer = viewer
-        self.with_viewer = with_viewer
+        self.canvas_only = canvas_only
         self.image = None
 
     def _repr_png_(self):
@@ -47,7 +67,7 @@ class NotebookScreenshot:
         """
         from imageio import imsave
 
-        self.image = self.viewer.screenshot(with_viewer=self.with_viewer)
+        self.image = self.viewer.screenshot(canvas_only=self.canvas_only)
         file_obj = BytesIO()
         imsave(file_obj, self.image, format='png')
         file_obj.seek(0)
